refactor: refatoração do padrão arquitetural incorreto implementado anteriormente
