+++ src/Application/DTOs/Authentication/GetAuthenticatedUserDto.cs
@@ -1,4 +1,4 @@
-namespace Application.Dtos.User
+namespace Application.DTOs.Authentication
 {
     public class GetAuthenticatedUserDto
     {
