+++ src/Application/Services/UserService.cs
@@ -217,7 +217,6 @@ namespace Application.Services
         public async Task<BaseResponse<GetUserDto>> DeactivateUserAsync(string username, int companyId)
         {
             User user = await _userRepository.GetUserByUsernameAsync(username);
-
             if (user == null)
             {
                 return new BaseResponse<GetUserDto>()
@@ -253,11 +252,6 @@ namespace Application.Services
                 Message = "Usuário foi desativado com sucesso.",
                 IsSuccess = true
             };
-        }
-
-        public async Task<BaseResponse<GetUserDto>> DeleteUserAsync(string username, int companyId)
-        {
-            throw new NotImplementedException();
         }
     }
 }
