+++ src/Infrastructure.Identity/ServicesExtension.cs
@@ -2,9 +2,9 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.DependencyInjection;
 using Domain.Entities;
-using Infrastructure.Data.DataContext;
-using Infrastructure.Identity.Contracts;
 using Infrastructure.Identity.Services;
+using Application.Contracts;
+using Infrastructure.Data.DataContext;
 
 namespace Infrastructure.Identity
 {
@@ -32,13 +32,13 @@ namespace Infrastructure.Identity
 
             // Set application login url path
             services.ConfigureApplicationCookie(options =>
-            {
-                options.Cookie.HttpOnly = true;
-                options.ExpireTimeSpan = TimeSpan.FromMinutes(30);
-                options.LoginPath = new PathString("/auth/login");
-                // options.AccessDeniedPath = new PathString(""); => create a custom page for this
-                options.SlidingExpiration = true;
-            });
+                {
+                    options.Cookie.HttpOnly = true;
+                    options.ExpireTimeSpan = TimeSpan.FromMinutes(30);
+                    options.LoginPath = new PathString("/auth/login");
+                    // options.AccessDeniedPath = new PathString(""); => create a custom page for this
+                    options.SlidingExpiration = true;
+                });
 
             return services;
         }
