+++ src/Presentation/Controllers/CustomerController.cs
@@ -8,6 +8,7 @@ using Presentation.ViewModels.Order;
 using Application.Dtos.Order;
 using Application.Dtos.User;
 using Presentation.Controllers.Common;
+using Application.DTOs.Authentication;
 
 namespace Presentation.Controllers
 {
