+++ src/Presentation/Controllers/MealController.cs
@@ -1,6 +1,7 @@
 using Application.Contracts;
 using Application.Dtos.Meal;
 using Application.Dtos.User;
+using Application.DTOs.Authentication;
 using Application.Responses;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
