+++ src/Presentation/Views/Authentication/Register.cshtml
@@ -1,4 +1,4 @@
-@model Infrastructure.Identity.Dtos.RegisterUserDto;
+@model Application.DTOs.Authentication.RegisterUserDto;
 
 @{
     ViewData["Title"] = "Cadastro";
